feat: add `createDiff`
This commit adds the `createDiff` function to create a report on any
function definition changes that have occured based on comparing
the current index.json with the new data.

We also remove the `.slice(1)` call previously used on `categories` in
`getPowerQueryFunctions`. This was to remove the table header rows.
The MS docs page has changed formatting making this no longer
necessary